Projeto Blog Pessoal/blogpessoal/src/main/java/com/generation/blogpessoal/repository/UsuarioRepository.java

package com.generation.blogpessoal.repository;

import java.util.List;
import java.util.Optional;

import org.springframework.data.jpa.repository.JpaRepository;
import org.springframework.stereotype.Repository;

import com.generation.blogpessoal.model.Usuario;

@Repository
public interface UsuarioRepository extends JpaRepository <Usuario, Long> {
	
	// Quando temos mais de uma possível resposta, usamos o Optional.
	public Optional <Usuario> findByUsuario(String usuario);

	public List<Usuario> findAllByNomeContainingIgnoreCase(String nome);
}